A few changes for IRIX 6.5/MIPSPro 7.2
--HG--
extra : convert_revision : svn%3A0fb12191-db4b-0410-8866-e78deaee71e5/trunk%402503

+++ mpqc/src/lib/util/ref/identity.cc
@@ -36,7 +36,7 @@
 void
 Identifier::print(ostream &o) const
 {
-  o << id;
+  o << (unsigned long) id;
 }
 
 ostream &
